fix bugs in resize cache limit. one can resize cache limit through monitor commond line.

+++ src/global.h
@@ -7,7 +7,7 @@ struct RuntimeSTAT
     unsigned int batchId;
     unsigned int userId;
     unsigned int traceId;
-    unsigned int startLBA;
+    unsigned long startLBA;
     unsigned int isWriteOnly;
 
     /** Runtime strategy refered parameter **/
